requirements.txt
streamlit==1.28.0
langchain==0.0.350
langchain-google-genai==0.0.8
python-dotenv==1.0.0
google-generativeai==0.3.2
tavily-python
requests
